+++ docs/Whats-New.md
@@ -178,6 +178,7 @@ Vanilla fixes:
 - Fix a bug where the sidebar could only contain up to 75 items on a strip (by ZivDero)
 - Fix a vanilla bug where harvesters would become permanently idle if they exhausted all resources to mine, even if new resources appeared (e.g. spawned by a Tiberium tree) (by JoyfulShush)
 - Fix a bug where if you started the game owning a Techno at its BuildLimit, it would not appear on your sidebar (by ZivDero)
+- Fix destroyed APCs sometimes not ejecting the infantry inside them when destroyed while moving (by JoyfulShush)
 
 :::